requirements.txt
openai
python-dotenv
slugify
